backend/requirements.txt
# Backend API + геообработка + парсинг (установка: pip install -r backend/requirements.txt)

# --- FastAPI ---
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.0.0
email-validator>=2.0.0

# --- HTTP / Open-Meteo ---
httpx>=0.24.0
requests>=2.31.0
openmeteo-requests>=1.2.0
requests-cache>=1.2.0
retry-requests>=2.0.0

# --- GeoTIFF / растр / фильтры ---
rasterio>=1.3.0
pyproj>=3.6.0
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.7.0
scipy>=1.10.0
scikit-image>=0.21.0
earthengine-api>=0.1.350

# --- Планировщик и парсинг ecodata.kz ---
APScheduler>=3.10.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
selenium>=4.15.0
pandas>=2.0.0
